@@ -1,7 +1,16 @@
-const path = require('path');
-const { createHash } = require('crypto');
-const { readFile, writeFile, unlink, appendFile } = require('fs/promises');
-const {
+import {
+  basename,
+  dirname,
+  extname,
+  sep,
+  posix,
+  isAbsolute,
+  resolve as _resolve,
+  relative as _relative
+} from 'node:path';
+import { createHash } from 'node:crypto';
+import { readFile, writeFile, unlink, appendFile } from 'node:fs/promises';
+import {
   getLogger,
   buildInjectCode,
   pluginName,
@@ -14,31 +23,30 @@ const {
   getRelativePath,
   getBuildId,
   validateNamedExport,
-  getPackageVersion
-} = require('./utils.js');
-const cssHandler = require('lightningcss');
-const camelCase = require('lodash/camelCase');
-const upperFirst = require('lodash/upperFirst');
-const BuildCache = require('./cache.js');
+  getPackageVersion,
+  getNormalPath,
+  emptySourceMap
+} from './utils.js';
+import { transform, bundle } from 'lightningcss';
+import { camelCase, upperFirst, pick, uniq, sortBy } from 'lodash-es';
 
 /**
  * buildCssModulesJs
- * @param {{fullPath: string; options: import('..').Options; digest: string; build: import('..').Build}} params
- * @returns {Promise<{resolveDir: string; js: string; css: string; originCss: string; exports: Record<string, string>}>}
+ * @param {{fullPath: string; options: import('../index.js').Options; digest: string; build: import('../index.js').Build}} params
+ * @returns {Promise<{resolveDir: string; js: string; css: string; originCss: string; exports: Record<string, import('lightningcss').CSSModuleExport>}>}
  */
 const buildCssModulesJs = async ({ fullPath, options, build }) => {
-  const cssFileName = path.basename(fullPath); // e.g. xxx.module.css?esbuild-css-modules-plugin-building
+  const cssFileName = basename(fullPath); // e.g. xxx.module.css?esbuild-css-modules-plugin-building
   const { buildId, relative, packageVersion, log } = build.context;
-  const resolveDir = path.dirname(fullPath);
-  const classPrefix =
-    path.basename(fullPath, path.extname(fullPath)).replace(/[^a-zA-Z0-9]/g, '-') + '__';
+  const resolveDir = dirname(fullPath);
+  const classPrefix = basename(fullPath, extname(fullPath)).replace(/[^a-zA-Z0-9]/g, '-') + '__';
   const versionString = packageVersion?.replace(/[^a-zA-Z0-9]/g, '') ?? '';
   const originCss = await readFile(fullPath);
-  const cssModulesOption = options.v2CssModulesOption || {};
+  const cssModulesOption = pick(options, ['pattern', 'dashedIndents']);
   const genTs = !!options.generateTsFile;
 
   /**
-   * @type {import('lightningcss').BundleOptions}
+   * @type {import('lightningcss').TransformOptions<any>}
    */
   const bundleConfig = {
     filename: relative(fullPath), // use relative path to keep hash stable in different machines
@@ -51,9 +59,10 @@ const buildCssModulesJs = async ({ fullPath, options, build }) => {
     },
     analyzeDependencies: false
   };
-  const { code, exports = {}, map } = cssHandler.transform(bundleConfig);
+  const { code, exports = {}, map } = transform(bundleConfig);
   let cssModulesContent = code.toString('utf-8');
 
+  /** @type {Record<string, string>} */
   const cssModulesJSON = {};
 
   Object.keys(exports)
@@ -80,7 +89,7 @@ const buildCssModulesJs = async ({ fullPath, options, build }) => {
   // fix path issue on Windows: https://github.com/indooorsman/esbuild-css-modules-plugin/issues/12
   const cssImportPath =
     './' +
-    cssFileName.split(path.sep).join(path.posix.sep).trim().replace(buildingCssSuffix, '') +
+    cssFileName.split(sep).join(posix.sep).trim().replace(buildingCssSuffix, '') +
     builtCssSuffix;
   // => ./xxx.module.css?esbuild-css-modules-plugin-built
   const importStatement = `import "${cssImportPath}";`;
@@ -98,6 +107,7 @@ export default new Proxy(${classNamesMapString}, {
   `
     : `export default ${classNamesMapString};`;
 
+  /** @type {string[]} */
   const namedExportsTs = [];
   const namedExportStatements = Object.entries(cssModulesJSON)
     .map(([camelCaseClassName, className]) => {
@@ -132,9 +142,9 @@ export default new Proxy(${classNamesMapString}, {
 
 /**
  * prepareBuild
- * @param {import('..').Build} build
- * @param {import('..').Options} options
- * @return {Promise<void>}
+ * @param {import('esbuild').PluginBuild} build
+ * @param {import('../index.js').Options} options
+ * @return {Promise<import('../index.js').Build>}
  */
 const prepareBuild = async (build, options) => {
   const buildId = await getBuildId(build);
@@ -143,8 +153,13 @@ const prepareBuild = async (build, options) => {
   const packageRoot = options.root;
   const buildRoot = getRootDir(build);
   const log = getLogger(build);
-  const relative = (to) => getRelativePath(build, to);
+  /** @param {string} to */
+  const relative = (to) => {
+    // @ts-ignore
+    return getRelativePath(build, to);
+  };
 
+  // @ts-ignore
   build.context = {
     buildId,
     buildRoot,
@@ -153,16 +168,18 @@ const prepareBuild = async (build, options) => {
     log,
     relative
   };
-  build.context.cache = new BuildCache(build);
 
   log(`root of this build(#${buildId}):`, buildRoot);
+
+  // @ts-ignore
+  return build;
 };
 
 /**
  * onResolveModulesCss
  * @description mark module(s).css as sideEffects and add namespace
  * @param {import('esbuild').OnResolveArgs} args
- * @param {import('..').Build} build
+ * @param {import('../index.js').Build} build
  * @returns {Promise<import('esbuild').OnResolveResult>}
  */
 const onResolveModulesCss = async (args, build) => {
@@ -196,34 +213,23 @@ const onResolveModulesCss = async (args, build) => {
 
 /**
  * onLoadModulesCss
- * @param {import('..').Build} build
- * @param {import('..').Options} options
+ * @param {import('../index.js').Build} build
+ * @param {import('../index.js').Options} options
  * @param {import('esbuild').OnLoadArgs} args
- * @return {(import('esbuild').OnLoadResult | null | undefined | Promise<import('esbuild').OnLoadResult | null | undefined>)}
+ * @return {(Promise<import('esbuild').OnLoadResult | null | undefined>)}
  */
 const onLoadModulesCss = async (build, options, args) => {
   const { path: maybeFullPath, pluginData = {} } = args;
-  const { buildRoot, log, cache } = build.context;
-  const absPath = path.isAbsolute(maybeFullPath)
-    ? maybeFullPath
-    : path.resolve(buildRoot, maybeFullPath);
+  const { buildRoot, log } = build.context;
+  const absPath = isAbsolute(maybeFullPath) ? maybeFullPath : _resolve(buildRoot, maybeFullPath);
   const rpath = pluginData.relativePathToBuildRoot;
 
   log(`loading ${rpath}${args.suffix}`);
 
-  const useCache = build.initialOptions.watch ?? options.watch ?? false;
-
-  useCache && log(`checking cache for`, rpath);
-  const cached = useCache && (await cache.get(absPath));
-  if (cached) {
-    log('return build cache for', rpath);
-    return cached;
-  }
-
   const hex = createHash('sha256').update(rpath).digest('hex');
   const digest = hex.slice(hex.length - 255, hex.length);
 
-  const { js, ts, resolveDir, css, exports, originCss } = await buildCssModulesJs({
+  const { js, resolveDir, css, exports } = await buildCssModulesJs({
     fullPath: absPath,
     options,
     digest,
@@ -240,13 +246,9 @@ const onLoadModulesCss = async (build, options, args) => {
       digest
     },
     contents: js,
+    /** @type {import('esbuild').Loader} */
     loader: 'js'
   };
-
-  if (useCache) {
-    await cache.set(absPath, result, originCss);
-    log(`add build result to cache for ${rpath}`);
-  }
 
   return result;
 };
@@ -254,7 +256,7 @@ const onLoadModulesCss = async (build, options, args) => {
 /**
  * onResolveBuiltModulesCss
  * @param {import('esbuild').OnResolveArgs} args
- * @param {import('..').Build} build
+ * @param {import('../index.js').Build} build
  * @returns {Promise<import('esbuild').OnResolveResult>}
  */
 const onResolveBuiltModulesCss = async (args, build) => {
@@ -281,14 +283,14 @@ const onResolveBuiltModulesCss = async (args, build) => {
 /**
  * onLoadBuiltModulesCss
  * @param {import('esbuild').OnLoadArgs} args
- * @param {import('..').Build} build
+ * @param {import('../index.js').Build} build
  * @returns {Promise<import('esbuild').OnLoadResult>}
  */
 const onLoadBuiltModulesCss = async ({ pluginData }, build) => {
   const { log, buildRoot } = build.context;
   const { css, relativePathToBuildRoot } = pluginData;
-  const absPath = path.resolve(buildRoot, relativePathToBuildRoot);
-  const resolveDir = path.dirname(absPath);
+  const absPath = _resolve(buildRoot, relativePathToBuildRoot);
+  const resolveDir = dirname(absPath);
   log('loading built css for', relativePathToBuildRoot);
 
   /**
@@ -307,8 +309,8 @@ const onLoadBuiltModulesCss = async ({ pluginData }, build) => {
 
 /**
  * onEnd
- * @param {import('..').Build} build
- * @param {import('..').Options} options
+ * @param {import('../index.js').Build} build
+ * @param {import('../index.js').Options} options
  * @param {import('esbuild').BuildResult} result
  */
 const onEnd = async (build, options, result) => {
@@ -319,7 +321,7 @@ const onEnd = async (build, options, result) => {
   if (options.inject) {
     const {
       charset = 'utf8',
-      outdir,
+      outdir = '.',
       sourceRoot,
       sourcemap,
       sourcesContent,
@@ -331,8 +333,9 @@ const onEnd = async (build, options, result) => {
       external,
       publicPath
     } = initialOptions;
-    const absOutdir = path.isAbsolute(outdir) ? outdir : path.resolve(buildRoot, outdir);
+    const absOutdir = isAbsolute(outdir) ? outdir : _resolve(buildRoot, outdir);
 
+    /** @param {string} css */
     const transformCss = async (css) => {
       const r = await esbuild.transform(css, {
         charset,
@@ -346,8 +349,18 @@ const onEnd = async (build, options, result) => {
       return r.code;
     };
 
+    /**
+     * buildJs
+     * @param {string} entryName
+     * @param {string} entryPath
+     * @param {string} jsCode
+     */
     const buildJs = async (entryName, entryPath, jsCode) => {
-      const r = (p) => path.relative(absOutdir, p).split(path.sep).join(path.posix.sep);
+      /**
+       * @param {string} p
+       * @returns {string}
+       */
+      const r = (p) => _relative(absOutdir, p).split(sep).join(posix.sep);
       const imports = `import "./${r(entryPath)}";\nexport * from "./${r(entryPath)}";`;
       if (sourcemap === 'external') {
         await appendFile(entryPath, `\n//# sourceMappingURL=${r(entryPath)}.map`, {
@@ -363,7 +376,7 @@ const onEnd = async (build, options, result) => {
         );
       }
       const tmpJsCode = `${imports}\n${jsCode}`;
-      const tmpJsPath = path.resolve(absOutdir, '.build.inject.js');
+      const tmpJsPath = _resolve(absOutdir, '.build.inject.js');
       await writeFile(tmpJsPath, tmpJsCode, { encoding: 'utf8' });
       await esbuild.build({
         charset,
@@ -389,22 +402,27 @@ const onEnd = async (build, options, result) => {
       await unlink(tmpJsPath);
     };
 
+    /** @type {string[]} */
     const cssContents = [];
-
+    /** @type {string[]} */
     let entriesArray = [];
     if (Array.isArray(entryPoints)) {
+      // @ts-ignore
       entriesArray = [...entryPoints];
     } else {
+      // @ts-ignore
       Object.keys(entryPoints)
         .sort()
         .forEach((k) => {
+          // @ts-ignore
           entriesArray.push(entryPoints[k]);
         });
     }
-    const entries = entriesArray.map((p) => (path.isAbsolute(p) ? p : path.resolve(buildRoot, p)));
+    const entries = entriesArray.map((p) => (isAbsolute(p) ? p : _resolve(buildRoot, p)));
 
     log('entries:', entries);
 
+    /** @type {string|null} */
     let entryToInject = null;
     const outputs = Object.keys(result.metafile?.outputs ?? []);
 
@@ -412,14 +430,14 @@ const onEnd = async (build, options, result) => {
       outputs.map(async (f) => {
         if (
           !entryToInject &&
-          result.metafile.outputs[f].entryPoint &&
-          entries.includes(path.resolve(buildRoot, result.metafile.outputs[f].entryPoint)) &&
-          ['.js', '.mjs', '.cjs'].includes(path.extname(f))
+          result.metafile?.outputs[f].entryPoint &&
+          entries.includes(_resolve(buildRoot, result.metafile?.outputs[f]?.entryPoint ?? '')) &&
+          ['.js', '.mjs', '.cjs'].includes(extname(f))
         ) {
-          entryToInject = path.resolve(buildRoot, f);
+          entryToInject = _resolve(buildRoot, f);
         }
-        if (path.extname(f) === '.css') {
-          const fullpath = path.resolve(buildRoot, f);
+        if (extname(f) === '.css') {
+          const fullpath = _resolve(buildRoot, f);
           const css = await readFile(fullpath, { encoding: 'utf8' });
           const transformed = await transformCss(css);
           cssContents.push(`${transformed}`);
@@ -428,8 +446,8 @@ const onEnd = async (build, options, result) => {
     );
 
     if (entryToInject && cssContents.length) {
-      log('inject css to', path.relative(buildRoot, entryToInject));
-      const entryName = path.basename(entryToInject, path.extname(entryToInject));
+      log('inject css to', _relative(buildRoot, entryToInject));
+      const entryName = basename(entryToInject, extname(entryToInject));
       const allCss = cssContents.join('\n');
       const container = typeof options.inject === 'string' ? options.inject : 'head';
       const injectedCode = buildInjectCode(container, allCss, buildId, options);
@@ -442,52 +460,166 @@ const onEnd = async (build, options, result) => {
 
 /**
  * setup
- * @param {import('..').Build} build
- * @param {import('..').Options} options
+ * @param {import('esbuild').PluginBuild} build
+ * @param {import('../index.js').Options} options
  * @returns {Promise<void>}
  */
 const setup = async (build, options) => {
-  await prepareBuild(build, options);
+  const patchedBuild = await prepareBuild(build, options);
   const modulesCssRegExp = getModulesCssRegExp(options);
   const builtModulesCssRegExp = getBuiltModulesCssRegExp(options);
 
+  const { log, relative, buildRoot, packageVersion } = patchedBuild.context;
+
+  /** @type {Map<string, import('lightningcss').TransformResult & {resolveDir: string}>} */
+  const resultMap = new Map();
+
+  patchedBuild.onLoad({ filter: /.*/, namespace: pluginNamespace }, async ({ path }) => {
+    const result = resultMap.get(path);
+    if (!result) {
+      return;
+    }
+    const { code, map, resolveDir } = result;
+    const codeWithSourcemap =
+      (code ? code.toString('utf8') : '') +
+      (map
+        ? `\n/*# sourceMappingURL=data:application/json;base64,${map.toString('base64')} */`
+        : '');
+    log(path, '==>\n', codeWithSourcemap);
+    return {
+      contents: codeWithSourcemap,
+      loader: patchedBuild.initialOptions.loader?.['.css'] ?? 'css',
+      resolveDir
+    };
+  });
+
+  patchedBuild.onResolve({ filter: new RegExp(`^${pluginNamespace}:`) }, async ({ path }) => {
+    return {
+      path: getNormalPath(path),
+      namespace: pluginNamespace
+    };
+  });
+
+  patchedBuild.onLoad({ filter: modulesCssRegExp }, async ({ path }) => {
+    const filename = relative(path);
+    const classPrefix =
+      basename(path, extname(path))
+        .replace(/[^a-zA-Z0-9]/g, '-')
+        .replace(/^\-*/, '') + '__';
+    const versionString = packageVersion?.replace(/[^a-zA-Z0-9]/g, '') ?? '';
+
+    const shouldBundle = !!options.bundle;
+
+    const buildFn = shouldBundle ? bundle : transform;
+
+    const result = buildFn({
+      filename: shouldBundle ? path : filename,
+      // @ts-ignore
+      code: shouldBundle ? undefined : await readFile(path),
+      cssModules: {
+        dashedIdents: options?.dashedIndents,
+        pattern: options?.pattern ?? `${classPrefix}[local]_[hash]${versionString}`
+      },
+      drafts: {
+        customMedia: true,
+        nesting: true
+      },
+      errorRecovery: true,
+      minify: false,
+      sourceMap: !!patchedBuild.initialOptions.sourcemap,
+      projectRoot: buildRoot,
+      targets: {
+        chrome: 112 << 16
+      }
+    });
+
+    resultMap.set(filename, { ...result, resolveDir: dirname(path) });
+
+    /** @type {string[]} */
+    const jsLines = [`import "${pluginNamespace}:${filename}";`];
+    const { localsConvention = 'camelCaseOnly' } = options;
+    const keepOrigin = !localsConvention?.endsWith('Only');
+    const useCamel = localsConvention?.startsWith('camelCase');
+    const usePascal = localsConvention?.startsWith('pascal');
+    /** @type {[string, string][]} */
+    const jsNames = [];
+    /** @type {[string, string][]} */
+    const originNames = [];
+    sortBy(Object.entries(result.exports ?? {}), '0').forEach(([origin, local]) => {
+      const jsName = useCamel
+        ? camelCase(origin)
+        : usePascal
+        ? upperFirst(camelCase(origin))
+        : camelCase(origin);
+      const composesNames = local.composes?.map?.((item) => item.name ?? '')?.join(' ');
+      const localName = `${composesNames ? composesNames + ' ' : ''}${local.name}`;
+      jsLines.push(`export const ${jsName} = "${localName}";`);
+      jsNames.push([jsName, jsName]);
+      if (keepOrigin && origin !== jsName) {
+        originNames.push([`"${origin}"`, jsName]);
+      }
+    });
+    jsLines.push(`export default {
+${uniq([...jsNames, ...originNames])
+  .map(([o, l]) => `  ${o}: ${l}`)
+  .join(',\n')}
+};`);
+
+    // https://github.com/evanw/esbuild/issues/1227
+    jsLines.push(`\n//${emptySourceMap}`);
+
+    const jsContents = jsLines.join('\n');
+    log(jsContents);
+
+    return {
+      contents: jsContents,
+      loader: 'js',
+      resolveDir: dirname(path),
+      watchFiles: [path],
+      pluginName: pluginName,
+      pluginData: {
+        hehe: 'hehe'
+      }
+    };
+  });
+
+  return;
+
   // resolve xxx.module.css to xxx.module.css?esbuild-css-modules-plugin-building
-  build.onResolve({ filter: modulesCssRegExp, namespace: 'file' }, (args) => {
-    return onResolveModulesCss(args, build);
+  patchedBuild.onResolve({ filter: modulesCssRegExp, namespace: 'file' }, (args) => {
+    return onResolveModulesCss(args, patchedBuild);
   });
 
   // load xxx.module.css?esbuild-css-modules-plugin-building
-  build.onLoad({ filter: modulesCssRegExp, namespace: pluginNamespace }, (args) => {
-    return onLoadModulesCss(build, options, args);
+  patchedBuild.onLoad({ filter: modulesCssRegExp, namespace: pluginNamespace }, (args) => {
+    return onLoadModulesCss(patchedBuild, options, args);
   });
 
   // resolve virtual path xxx.module.css?esbuild-css-modules-plugin-built
-  build.onResolve(
+  patchedBuild.onResolve(
     {
       filter: builtModulesCssRegExp,
       namespace: pluginNamespace
     },
     (args) => {
-      return onResolveBuiltModulesCss(args, build);
+      return onResolveBuiltModulesCss(args, patchedBuild);
     }
   );
 
   // load virtual path xxx.module.css?esbuild-css-modules-plugin-built
-  build.onLoad(
+  patchedBuild.onLoad(
     {
       filter: builtModulesCssRegExp,
       namespace: pluginNamespace
     },
     (args) => {
-      return onLoadBuiltModulesCss(args, build);
+      return onLoadBuiltModulesCss(args, patchedBuild);
     }
   );
 
-  build.onEnd(async (result) => {
-    onEnd(build, options, result);
+  patchedBuild.onEnd(async (result) => {
+    onEnd(patchedBuild, options, result);
   });
 };
 
-module.exports = {
-  setup
-};
+export { setup };
