@@ -1,8 +1,10 @@
-const path = require('path');
-const { createHash } = require('crypto');
-const { readFile } = require('fs/promises');
-const fs = require('fs');
-const pluginName = require('../package.json').name.toLowerCase();
+import { isAbsolute, resolve, sep, relative } from 'node:path';
+import { createHash } from 'node:crypto';
+import { readFile } from 'node:fs/promises';
+import { accessSync, constants } from 'node:fs';
+
+const pluginName = `esbuild-css-modules-plugin`;
+
 const pluginNamespace = `${pluginName}-namespace`;
 const buildingCssSuffix = `?${pluginName}-building`;
 const builtCssSuffix = `?${pluginName}-built`;
@@ -10,7 +12,7 @@ const builtCssSuffixRegExp = builtCssSuffix.replace('?', '\\?').replace(/\-/g, '
 
 /**
  * getModulesCssRegExp
- * @param {import('..').Options} options
+ * @param {import('../index.js').Options} options
  * @returns {RegExp}
  */
 const getModulesCssRegExp = (options) => {
@@ -19,7 +21,7 @@ const getModulesCssRegExp = (options) => {
 
 /**
  * getBuiltModulesCssRegExp
- * @param {import('..').Options} options
+ * @param {import('../index.js').Options} options
  * @returns {RegExp}
  */
 const getBuiltModulesCssRegExp = (options) => {
@@ -32,7 +34,7 @@ const getBuiltModulesCssRegExp = (options) => {
 
 /**
  * getLogger
- * @param {import('..').Build} build
+ * @param {import('esbuild').PluginBuild} build
  * @returns {(...args: any[]) => void}
  */
 const getLogger = (build) => {
@@ -50,7 +52,7 @@ const getLogger = (build) => {
  * @param {string} injectToSelector
  * @param {string} css
  * @param {string} digest
- * @param {import('..').Options} options
+ * @param {import('../index.js').Options} options
  * @returns {string}
  */
 const buildInjectCode = (injectToSelector = 'head', css, digest, options) => {
@@ -94,26 +96,26 @@ const buildInjectCode = (injectToSelector = 'head', css, digest, options) => {
 
 /**
  * getRootDir
- * @param {import('..').Build} build
+ * @param {import('esbuild').PluginBuild} build
  * @returns {string}
  */
 const getRootDir = (build) => {
   const { absWorkingDir } = build.initialOptions;
   const abs = absWorkingDir ? absWorkingDir : process.cwd();
-  const rootDir = path.isAbsolute(abs) ? abs : path.resolve(process.cwd(), abs);
+  const rootDir = isAbsolute(abs) ? abs : resolve(process.cwd(), abs);
   return rootDir;
 };
 
 /**
  * getPackageInfo
- * @param {import('..').Build} build
+ * @param {import('esbuild').PluginBuild} build
  * @returns {{name: string; version: string;}}
  */
 const getPackageInfo = (build) => {
   const rootDir = getRootDir(build);
-  const packageJsonFile = path.resolve(rootDir, './package.json');
+  const packageJsonFile = resolve(rootDir, './package.json');
   try {
-    fs.accessSync(packageJsonFile, fs.constants.R_OK);
+    accessSync(packageJsonFile, constants.R_OK);
     return require(packageJsonFile);
   } catch (error) {
     return { name: '', version: '' };
@@ -122,7 +124,7 @@ const getPackageInfo = (build) => {
 
 /**
  * getPackageVersion
- * @param {import('..').Build} build
+ * @param {import('esbuild').PluginBuild} build
  * @returns {string}
  */
 const getPackageVersion = (build) => {
@@ -132,35 +134,38 @@ const getPackageVersion = (build) => {
 /**
  * getRelativePath
  * @description get relative path from build root
- * @param {import('..').Build} build
+ * @param {import('../index.js').Build} build
  * @param {string} to
  * @returns {string}
  */
 const getRelativePath = (build, to) => {
-  if (!path.isAbsolute(to)) {
-    return to.startsWith('.') ? to : `.${path.sep}${to}`;
+  if (!isAbsolute(to)) {
+    return to.startsWith('.') ? to : `.${sep}${to}`;
   }
   const root = build.context?.buildRoot ?? getRootDir(build);
-  return `.${path.sep}${path.relative(root, to)}`;
+  return `.${sep}${relative(root, to)}`;
 };
 
 /**
  * getBuildId
  * @description buildId should be stable so that the hash of output files are stable
- * @param {import('..').Build} build
+ * @param {import('esbuild').PluginBuild} build
  * @returns {Promise<string>}
  */
 const getBuildId = async (build) => {
   const { entryPoints } = build.initialOptions;
   const buildRoot = getRootDir(build);
   const { version: packageVersion, name: packageName } = getPackageInfo(build);
+  /** @type {string[]} */
   let entries = [];
   if (Array.isArray(entryPoints)) {
+    // @ts-ignore
     entries = [...entryPoints];
   } else {
     Object.keys(entryPoints ?? {})
       .sort()
       .forEach((k) => {
+        // @ts-ignore
         entries.push(entryPoints[k]);
       });
   }
@@ -172,7 +177,7 @@ const getBuildId = async (build) => {
           if (!p) {
             return Promise.resolve('');
           }
-          const absPath = path.isAbsolute(p) ? p : path.resolve(buildRoot, p);
+          const absPath = isAbsolute(p) ? p : resolve(buildRoot, p);
           return readFile(absPath, { encoding: 'utf8' }).catch(() => '');
         })
       )
@@ -237,7 +242,19 @@ const validateNamedExport = (name) => {
   return !jsKeywords.includes(name);
 };
 
-module.exports = {
+/**
+ * getNormalPath
+ * @param {string} pathWithNamespace
+ * @returns string
+ */
+const getNormalPath = (pathWithNamespace) => {
+  return pathWithNamespace.replace(new RegExp(`^${pluginNamespace}:`), '');
+};
+
+const emptySourceMap =
+  '# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIiJdLCJtYXBwaW5ncyI6IkEifQ==';
+
+export {
   pluginName,
   pluginNamespace,
   getLogger,
@@ -251,5 +268,7 @@ module.exports = {
   getBuildId,
   validateNamedExport,
   getPackageInfo,
-  getPackageVersion
+  getPackageVersion,
+  getNormalPath,
+  emptySourceMap
 };
